+++ CHANGELOG.md
@@ -5,8 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## [unreleased]
+
+### Added
+
+- Add `OtcMaker` contract.
 
 ## [1.4.3] - 2023-04-11
 ### Added
@@ -14,7 +17,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
 - `PerpPortal.getSqrtMarkTwapX96(address baseToken, uint32 twapInterval)` will be deprecated at later releases. Suggest to use `PerpPortal.getSqrtMarketTwapX96()` instead.
-
 
 ## [1.4.2] - 2023-04-10
 - Support Curie v2.6.0